Updates for documentation, READMEs, and code format

+++ examples/get_started_pytorch.py
@@ -41,8 +41,8 @@ class Net(nn.Module):
 
 # Step 1a: Swap axes to PyTorch's NCHW format
 
-x_train = np.swapaxes(x_train, 1, 3)
-x_test = np.swapaxes(x_test, 1, 3)
+x_train = np.swapaxes(x_train, 1, 3).astype(np.float32)
+x_test = np.swapaxes(x_test, 1, 3).astype(np.float32)
 
 # Step 2: Create the model
